+++ NEWS.md
@@ -2,7 +2,8 @@
 
 - Bug fix for `RImissingP()`, which has been broken a while.
 - Experimental new function `RIinfitKfold()` for cross-validating item fit using `rsample::vfold_cv()`; "V-fold cross-validation (also known as k-fold cross-validation) randomly splits the data into V groups of roughly equal size (called "folds"). A resample of the analysis data consists of V-1 of the folds", see <https://rsample.tidymodels.org/reference/vfold_cv.html>
-  - this has not yet been tested so there are no recommendations for how or when to use it
+  - this has not yet been tested so there are no recommendations for how or when to use it.
+- Slight speedup (~20%) for `RIgetResidCor()` by changing the accelerator used by `mirt()` to SQUAREM and avoiding the use of `sink()`.
 
 # easyRasch 0.3.6.2 (2025-04-30)